Update CODE_OF_CONDUCT.md
Added 'such as but not limited to' to give context to the list.  Added unconscious beings.  "CoC ranting" is too vague and actual issues with the CoC (wording issues, grammar errors, vagueness, etc should probably be accepted so long as they stay apolitical and relevant).  Capitalized lists.  Removed "on anything non-project related", your views on the project shouldn't matter either